Support the use case of two different users uploading a same map.
This changes the way we're saving maps in the database by quite a bit.
Now there's a new table, that serves as a join table in a many-to-many
relationship between maps table and users table.

This might complicate some queries we might have in the future though
(eg. getting a map along with information on who uploaded it), but I
don't see a better way to implement it.

Also, this commit removes the `filename` and `modified_time` time of the
map files, since they're largerly irrelevant to us.