Introduce the concept of ownership and visibility to our map system. (#408)
* Introduce the concept of ownership and visibility to our map system.

* Clear the old lobby preferences before getting them.

Not clearing the old lobby preferences can cause problems when logging
in with two accounts (old ones can be transferred to the new one if the
new one doesn't have its own lobby preferences).

* Check if the selected map exists in the recent maps list before returning
lobby preferences.

This can happen in certain cases when maps get deleted. Recent maps list
filters out all non-existing maps, so selected map needs to check if
it's not the map that got filtered out.

* Add throttling to the maps api.

* Support the use case of two different users uploading a same map.

This changes the way we're saving maps in the database by quite a bit.
Now there's a new table, that serves as a join table in a many-to-many
relationship between maps table and users table.

This might complicate some queries we might have in the future though
(eg. getting a map along with information on who uploaded it), but I
don't see a better way to implement it.

Also, this commit removes the `filename` and `modified_time` time of the
map files, since they're largerly irrelevant to us.

* Run the transactionFn only if a new map is added.