Add a standalone map browser available from activity bar.
So far, the only way to get to the map browser was through 'create lobby'
workflow, which could be used to select a map to host. Now, both the
server and local map browsers have been extended to be able to be used
as a standalone browsers, which will in future allow for more actions on
the maps than the create lobby workflow will allow.