Replace the delete map API with a soft delete.
After some discussion, it was decided that maps should probably never be
deleted. Since a lot of different parts of the application (recent maps
list, matchmaking map pools, replays info, etc.) relies on map API,
deleting a map would leave those parts of the application in some kind
of a broken state that would need to be handled.

To avoid that, maps can't be deleted at all now, although we still
support the use-case of users removing a map from their list of private
maps, or admins removing maps from the official and/or public lists.
Those maps will have a `removed_at` column set to the date of their
removal and they will *only* be filtered from the listing maps API, for
now. In future, `removed_at` column can be used to filter maps from
other parts of the application if there's a need for it.