Keep the init data as part of the activity overlay history.
In the previous implementation of activity overlay state, there was a
bug where `initData` would only be kept for the latest opened activity.
This could cause issues when going through the activity overlay history
and expecting different data. That problem has now been fixed and also
used to provide `title` value as a prop to the server maps browser.