Fix an annoying issue where some libraries might require `fs` in browsers
I honestly have no idea which library is even doing this, but since
it started to be an issue from the last commit where
`react-transition-group` was added, I will assume it is that one. Or one
of *its* dependencies.

In any case, it's really annoying and this "fix" doesn't seem optimal to
me; I'm not even sure if it didn't break anything else (from my limited
testing, it didn't). See this github issue for a bit more info:
https://github.com/tailwindcss/tailwindcss/issues/1129