Upgrade deps. (#426)
* Upgrade deps.

* Add `request` dependency explicitly.

Not sure how this used to work, but now webpack is throwing that it
can't find this dependency so we're installing it explicitly.

* Fix ambiguous composition in css modules.

Not sure how this used to work, but now it throws an error.

* Update `css-loader` options.

* Import `upload` function only in Electron environment.

Reverts commit fbfb355.
Previously, I thought it was some library that's requiring the `fs`
module, and it turns out it was me after all. Oops.

* On second thought, rever the `style-loader` dependency as it has a
pretty bug

Need to wait until this issue is resolved before updating this:
https://github.com/webpack-contrib/style-loader/issues/450

Not a big deal, only took me 2 hours to figure out what's going wrong
here. The day we completely move to `styled-components` can't come
any sooner...

* Update deprecated Electron APIs.

* Fix prettier lint errors.

* Add a temporary fix for overriding default button styles.

* Import Electron `screen` API from the main process.