Fix a bug when updating game subtype field in lobby creation.
Game subtype wouldn't get updated if the creaty lobby overlay was
initialized with a map that couldn't support previously set subtype.

Also, this commit slightly changes the logic in which we set game
subtype. Now, every time the game type or the selected map changes, we
set the game subtype back to the default value (which is equal teams in
top vs bottom games and 2 teams in team melee games).