Fix a bug when auto-focusing lobby name field in lobby creation overlay.
We used to auto-focus the lobby name field at the same time as getting
the lobby preferences. Which meant that if the timer for the
auto-focusing expired before we get the lobby preferences, the
reference to the `input` field was invalid because it was not rendered
yet (since we render the loader until lobby preferences fetching is
done).