Basic support for newer versions (#435)
This is still quite incomplete but putting it up here because replays and playing alone vs ai work.
(At least for me)

Multiplayer/observing don't work.

A too big overhaul of the existing code wasn't necessary, once the game has
been launched it functions mostly the same. Some struct layouts are a different (And obviously
the addresses as well) but that's mostly it.

Forge isn't used here, though other parts of code call directly to forge so it was changed
to have hacky support for non-forge windows. Probably should refactor it so that Bw1161 calls
to forge and BwScr does something else, and everyone goes through that interface.

Most of the changes were during launch:

1) The JS code of app which verifies the version will now recognize if the selected starcraft
directory has "x86/starcraft.exe" in it and uses that to decide if we launch 1161 mode or not
(Game dll relies on '-launch' cmdline argument being passed/not for detecting this).

The code was written pretty quickly so it could likely be organized a lot better, I just didn't
want to focus on it too much :p

2) The App's C++ side got a alternate injection implementation which launches the game as a
process to be debugged, uses windows apis to debug it a bit in order to get out DLL injected
right before Windows calls TLS callbacks of the exe. The code is messy, maybe relies a bit on Windows internals (doesn't
work on Wine, should work on Win7+) and I really hate it and wish that I never have to touch
it again.

Hopefully it won't cause any issues.

The api won't also fit 1:1 with the older injection code, most notably inject+resume are a single
operation with this newer method, the explicit `process.resume()` is only meaningful with 1.16.1.
(It could be developed further to separate inject and resume from each other but I really would
rather not)

3) Once the process is launched, we still need to do a bit of.. trickery in order to be able to patch
it reliably. That part is done in a black-box dll sb_init.dll. Don't worry too much about how 
it works =) Once sb_init.dll is ready it calls a callback that was provided to it and we can
patch the exe.

4) After all that, we still need to figure out what memory addresses are used by the current
exe for variables/functions. Conveniently, I happen to have a library called `samase_scarf` which
takes the exe binary data as input and is able to figure out those important addresses.

It is quite complex project and I really should document it, but shieldbattery's side is nice and clean,
needing only few lines of code per variable.

After that everything mostly just works. There were minor differences at few places, e.g. 1.16.1
allows 255 as invalid storm player id but now it has to be 0xffffffff or else the game crashes

# Additional things

The server has a 30 second launch timeout specified at `server/lib/wsapi/lobbies.js`. It may
be worth it to increase that when trying this (At least it would have been necessary for this PC,
but I have a locally patched it very high anyway since it gets in the way of debugging launch issues)

Since debugging this is quite tough compared to 1.16.1, and there can be unexpected crashes on
other systems, I added a simple crash handler that writes a minidump to logs folder on crash.
Debugging those can be often easier than live debugging (Still not that simple).
The crash message needs improvement though so that people know where to find the crash dump and
such. 1.16.1 also benefits from that if it ever crashes. Even still should probably explicitly
block the usual error reporter as it has sent minidumps on some versions for just assertion
failures.

* Injection patches

* Able to play alone

* Fix issues

* Integrate forge more with scr

* Minor dependency update

* Make SCR support be behind a flag

* Add logging callback to the native process code

Implementation has to make sure JS objects are only used on the JS thread
(which is not the same thread what process creation is done on)

* Remove an unnecessary commented out debug message