Update for 1.23.3 (#439)
Two different things broke:

1) SNP can no longer be replaced by just writing over global variables.
Now we hook a function that provides SNPs to BW instead.

Predicting future more than that is probably not possible, we'll see if
there are more SNP changes in future

2) The lobby dialog's vtable got a new function, pushing one we have to
implement to be located at a different offset. This could have been made
more resilient by just filling the vtable with this function (as no other
functions are expected to be called), or having more analysis to figure
out the correct offset instead of just hardcoding it.