Move the maps page state to the component from the store.
There wasn't much sense to keep this state in the store in the first
place. Now, hopefully, there will be less chance of bugs happening with
page desyncing when applying different filters.

Also, we don't rely on server returning us an empty set anymore, to stop
fetching more maps. We get a `total` count of maps from the server on
each request, no reason not to use it.