Move the z-index of snackbar onto its container.
Since the container of snackbar is an element with fixed position, it
creates its own stacking context; which in turn means that snackbar
doesn't get displayed on top of an activity's scrim for example. This
commit makes the container itself be the one that its z-index is used in
determining correct positioning.