Turn the downgrade feature off as it's not necessary anymore.
For now we're assuming that every SC:R version is supported since we're
updating our offsets for each new patch dynamically thanks to neive's
magic.

If that assumption ever breaks, we can turn the downgrade system on at
the flick of a switch (will probably need to be updated a bit for SC:R,
but not by much).

As for the 1.16.1 version of BW, we still support it fine, although now
we don't support all the intermediate versions between 1.16.1 and SC:R.
Which is fine.