Fix babel setup to work with local dependencies.
This fixes things such that the server can run without installing any of
the other projects' dependencies (e.g. the root project). Before this,
the babel-register-hook was causing it to load things out of the root
project a lot of the time.

This setup was also broken for packaged electron apps, so it had to be
fixed there as well. Note that the fix is slightly different there, as
Babel didn't seem to want to load the json file out of the ASAR (despite
it being there). Thankfully you can just pass the options right to
babel/register, so I did that. There were also a few things that had
changed in electron-builder, so I had to update the yml configs and
commands to make things work again. We should probably consider using
webpack to build the app's main script bundle as well, now that it has a
target for it. Would make things simpler to not be using babel on
packaged builds (and maybe mean we don't need app/common any more?).

Lastly, I touched up the signing stuff in our game build script, which
was broken since the Rust changes and didn't include signing for
sb_init.dll. I *think* it's okay to sign that, and added it, but if not
we can just roll that piece back.