Refactor the email verification code to use less information.
To verify an email and/or send a verification email a user needs to be
logged in, so now we're simply reading the `userId` from the session and
we fetch their email from the database. This allows us to send less
information to the client.

Note: there can arise a potential confusion if a user clicks on "Verify
email" button they got in their email while being logged in with a
*different* account than what their email was for. They will get a
message that the code is invalid, which might be a bit confusing if they
don't realize that they have to be logged in with an account that they
want to verify. However, I think this is a particularly rare use-case,
that it's not worth doing anything special about.

Also, this PR fixes a nasty bug that occurred by not awaiting a function
that updates all sessions.