Add CSP to the Electron+Web app.
Electron:
This required some shenanigans since we can't control the headers on a
file:// response (how we load the HTML in electron), and the stuff we
can set from a meta tag is rather limited. I wanted to use the nonce
approach, since that seemed to be the only approach that would support
styled-components' style tags.

Fortunately Electron has a thing for registering custom protocols and
returning content to requests to them. I made one for shieldbattery://
and return the index.html contents to it (with the proper URLs plugged
in, a bonus feature we get now! :) ), along with the desired CSP
headers.

This triggers some other settings to flip in Electron (sandbox, for some
reason), so I had to change those back, but otherwise it seems to
"work". And by "work", I mean we have a bunch of stuff I have to fix in
our codebase to make this work still.

I had to remove thenify from client since it uses eval. I removed it
from app as well since better alternative APIs existed for everything
that used it.

I also had to add "proper" CORS support to the server, since Electron
cares about that from non-file:// origins. As a benefit of using a
non-file:// protocol, we can now remove our faked origin code and
instead just check for shieldbattery:// on the server.

Note: In development we have to allow 'unsafe-eval' for hot reloading
to work. This (unfortunately) means Electron will warn about CSP
still, and we can't turn that warning off without turning off all
other security warnings, so I've left it on. You can safely ignore it.
*But* this also means if you check in code that uses `eval` you may
not notice until running things in production mode. I don't expect
this to be a common case, but something to look for when choosing
deps.