Adjust transitions/styling of Select/Menu components.
The Select component now only scales vertically, which makes more sense
for what is happening positionally. I also adjusted the timings/curves
to make things feel snappier and more consisten. I also bumped up the
shadow a bit because it wasn't very noticeable on our dark theme
(shadows in general really aren't as effective, need to figure out some
kind of treatment to divide things better).

For the menu I changed how it calculates sizing, such that it leaves an
item half-cut-off if there's a scroll-bar, which should make it more
obvious to users that there's more options to be found.