Implement a carousel component.
This component can be used as a way to display a list of items, for
example maps. Includes the support for dynamically loading the items
as well, so it can be used as an alternative way to do pagination.

Also, we update the infinite-scroll-list component in this PR, which now
exposes a hasMoreData property which can be sent to it to indicate if
there's more data to load. This makes users of this component in charge
of deciding when to stop trying to load more items.