Implement a carousel component. (#455)
* Implement a carousel component.

This component can be used as a way to display a list of items, for
example maps. Includes the support for dynamically loading the items
as well, so it can be used as an alternative way to do pagination.

Also, we update the infinite-scroll-list component in this PR, which now
exposes a hasMoreData property which can be sent to it to indicate if
there's more data to load. This makes users of this component in charge
of deciding when to stop trying to load more items.

* Add a gradient to items when they're scrolled.

Also fix a bug in carousel test by sending an actual array as items,
instead of whatever immutable's Range() returns, since it was screwing
the dynamically load of carousel items.

* Combine the carousel gradients into a single one with multiple stops.

* Add some scroll leeway when deciding whether to scroll until the
beginning/end of the list.

Co-authored-by: Travis Collins <travis@tec27.com>