Remove the race enum name. If you ran this migration, check out the previous commit and undo it.
For a common setup, that migration would have failed due to not having
superuser privileges, so I removed the contents of it instead of writing
a rollback. The app code was also not properly updated to handle the new
values (there are a *lot* of places that use single-letter race codes,
including our Rust codebase). Renaming the enum seemed to add more
conversions than it removed (I didn't see any conversions it removed),
so I don't think there's a good reason to rename the enum values after
all.