Refactor the game loading code.
This PR takes the code that was inside a lobby service and pulls it into
its own thing. Now we have two new pieces to deal with the game loading:
- Game coordinator
- games API

Game coordinator is a service which is (currently) in charge of keeping
track of games that are loading, as well as setting up the routes for
all players. In the (near) future, it will also be used to start
tracking results for each loaded game. Another thing that might be
useful to have is to track the currently active game for all users, even
for single-player games (like watching a replay, or demo-ing a map).
This would allow us to display the information next to the user's name
on the client-side if they're currently in the game, and what exactly is
the context of their game.

Games API is currently serving a purpose to allow clients to report when
their game has successfully loaded, or when it has failed loading. In
(near) future, it will be used to report game results as well. Also,
even though the API is currently called from the client, it might make
more sense to call the API directly from the game in future.