Add a game activity nav entry and use it for lobbies and matchmaking.
The game activity nav entry displays the information about loading and
active games. This commit also refactors a bunch of stuff, mostly
regarding our previous version of active game entry, and it updates the
lobby stuff to work similarly to the matchmaking stuff, i.e. by using
dedicated routes to display a loading screen and an active game screen.

Currently, the lobbies don't have an active game screen implemented, as
it needs to be designed first. Maybe we can just reuse the one we're
using in matchmaking for now?

Also, replays are not hooked up to this new system at all for now, but
I'll open a separate issue for that.