Generate a smaller sized image for all maps and use it as a thumbnail. (#499)
* Generate a smaller sized image for all maps and use it as a thumbnail.

This PR assumes we're only gonna generate two images for each map, so
it doesn't bother making the code too generic to work for N amount of
images. That seems like a reasonable assumption for now. If it ever
breaks, some kind of an array format will need to be used to work with
the image URLs, and their saving path might need to be rethought.

* Address PR comments.

* Add some things that I forgot.