Create a MapImage component that should be used as a primitive for displaying map images. (#501)
* Create a MapImage component that should be used as a primitive for
displaying map images.

Also uses it throughout the app. There are two components that are a
level higher than MapImage component, which are MapThumbnail and
MapPreview. They can be used in places where some common functionality
is required to make things easier (though MapPreview is pretty bare
atm).

This PR also updates the naming scheme of the map image names and URLs
to be a bit more cleaner and consistent. They are also saved in the same
folder now.

* Address PR comments.