Allow the slider component to hide ticks on a track through props.
When there are a lot of steps in a slider, say a 100, showing ticks look
pretty bad, so we allow users of the component to hide them if they want.
Might be better to dynamically decide whether to show the ticks based on
the slider width and the ticks count, but it's unclear what's the best
formula for that, so we're going with this for now.