Attempt fixing the laggy sliders by lowering the transition duration.
This fix feels a bit iffy since it was done on trial and error basis, so
it's even questionable how it will behave on other user's computers. It
definitely should make things better, so it's not for nothing, but there
might still be a bit of a lagg when pressing down a key for long or
dragging pretty fast with a mouse.

Fixes #478