Inline the sending of the account verification email.
I have no idea why I pulled this out in a separate function before when
all it did is call a different function. Anyway, now we have one less
folder to worry about.

Also, the sending of emails is not awaited anymore; we just log the
error if the email sending fails.