Add typescript support. Only on the server for now.
The game result code (and I'm sure a bunch of other already-written
stuff) can benefit pretty heavily from having a real type system behind
it, and will make it way easier to track down problems in the code I'm
writing, as well as making VSCode generally more useful. At this point I
have no plans to try any kind of mass-migration to TS in the whole
codebase, I'm just going to do things as I modify them/as they are
useful.

This adds another step to linting, which is to run `tsc` over the
codebase and typecheck things. Otherwise, we still use babel to build at
runtime, which for TS basically just throws away the types and doesn't
check them.

We can probably pretty easily turn this on for the client in the near
future, but I mainly want the benefits for the server atm. As a side
effect of this I had to use a different eslint parser which picked up on
different things (and failed on certain things, namely the
never-gonna-happen ::bind operator that I've been meaning to remove from
our code for sometime now, so that's gone now). I also bumped the ESLint
React version which triggered other warnings that we need to fix
sometime soon.