Add SC:R settings.
The SC:R settings are dealt with in a similar manner to our own local
settings. We watch the SC:R settings file and merge settings directly to
it. Also, we don't try to be smart and detecting which exact settings
have changed, and instead just simply save both the local and SC:R
settings every time.