Add SC:R settings. (#524)
* Add SC:R settings.

The SC:R settings are dealt with in a similar manner to our own local
settings. We watch the SC:R settings file and merge settings directly to
it. Also, we don't try to be smart and detecting which exact settings
have changed, and instead just simply save both the local and SC:R
settings every time.

* Address PR comments

* Update the slider design to the new spec

* Cleanup the naming/ordering/layout of the settings.

* Save the SC:R settings in our own settings file.

* Setup a watcher on SC:R settings as well, so we can merge our own settings with them instead of doing a full overwrite.

* Standardize setting names so theres no conflict between 1.16.1 and remastered versions.

* Add a quick wrapper around the TextField component to handle numbers in forms.

* Add a bit of padding between various elements.

* Use async IPC for setting overwriting.

* Address PR comments.

* Update a comment.