Use the precedence boost to fix the raised button styling.
See this comment for more info:
https://github.com/styled-components/styled-components/issues/3244#issuecomment-687676703

From what I can tell reading this comment, we pretty much don't ever
want to use the `&` to refer to the static component and instead `&&`
should be used to refer to the dynamically generated class for a
specific instance? The reason being is that referring to the static
component can lead to unexpected results when there are multiple
instances of that component rendered at the same time, and trying to
style certain instances differently.

This commit fixes just the styling of the raised button since that's
where I noticed this happening first, but it's very likely there are
other places where this change in how styled-components work leads to
unexpected styling. Those should be fixed when noticed.