Send the game results to the server on game end. (#539)
This sets things up so we can make this request from the Rust code, as the game finishes. We make 3 attempts, if all 3 fail, the expectation is that the app will retry at some point in the future (but this is NOT implemented yet as of this PR). The app can know to do this by the fact that the game exits without calling /game/resultSent (and my plan is to always submit some sort of results if that occurs, to try to give us fewer result timeouts in the case of weird issues).

APM calculation is also not included in this PR, for now everybody is registered as having 0 APM.

Note that there's a number of random Rust reformatting changes in here, as VSCode really wanted to do them. I adjusted our .editorconfig to preserve most of the existing formatting.