Add user popover menu in userlist (#534)
Fixes #269

I copied self-profile-overlay.jsx and only made slight changes for user-overlay.jsx
I also added a hover effect to the users in the userlist to make it more clear that it can be clicked. I think it looks a bit odd, not entirely sure how it should be styled.