Add new typography scale styles.
I attempted to improve on the older version by:

- Exporting a lowercase css-only version of each style, for cases where
you want the text style but also want to apply custom spacing/color/etc.
to it

- Exporting an UpperCase styled div, for when you want to style some
text exactly that way and it will be sized/positioned by some container
(e.g. in a flex layout).

It may make more sense to use semantic elements for the UpperCase
versions and set their margins explicitly (to override the reset
defaults), but I'm not entirely sure yet, so for now I've avoided
determining those values.