Add a way for admins to trigger map image regeneration.
This required us to have a way to read files in the file store, so in
the process of adding that I TS-ified the FileStore code to make it
easier to maintain/keep in sync.

The API reads the file out of the store, writes it to a temporary local
file, triggers the map-parse-worker again (only for the image output),
writes the images, then deletes the temp file.

If we need to regenerate user map images at some point (probable), we'll
probably need some way for admins to browse all the uploaded maps. This
would probably be nice to have anyway, but I didn't want to spend time
implementing that UI. So we'll do it later :)

bw-chk was upgraded as part of this (since that's why we wanted to be
able to regenerate images), but the current published version was broken
for our babel setup, so I fixed that and pegged us to that commit for
now.

This also moves the map file code out of common/. We were only using the
file code in one place (in the map-parse-worker) and our web code was
blindly importing out of this common file, which meant it was blindly
importing bw-chk, which meant that bw-chk had to be usable in a browser
(which it is not, at least not with our current version of webpack).
Moving this probably has some good bundle size implications as well, but
who knows!