Upgrade to Webpack 5.
I was noticing that our babel-loader was ignoring the `babelrc: false`
option (and maybe the other ones too?) so my guess is that maybe it
doesn't work with babel 4 properly any more, we'll see I guess :)

This was mostly painless, just required removing a few options that had
long been removed and/or changed, and moving to a custom wrapper for
webpack-dev-middleware (koa-webpack pulls in a version of
webpack-hot-client blindly, and that tries to import something out of a
private file in Webpack that has since been renamed). We weren't using
much of koa-webpack anyway, so the resulting replacement was pretty
simple.