Convert app.js -> TypeScript, add necessary typings, start using tsyringe.
For now tsyringe is only injecting things related to WebsocketServer,
and WebsocketServer creates a child container to create it's socket
groups with the right session lookup. In the future I think we could
make a root class for App so that we can easily inject all the things
that need to depend on WebsocketServer (e.g. MatchmakingStatus) without
having to make the resolve call manually.

tsyringe is also only available in the server code, not sure we'd want
to push it to the client (since it needs a polyfill that's a bit large)
and I'm not sure how useful it would be there. Perhaps once we have more
TS on that side of things too.