Make 1.16.1 init_game_data hook match SC:R
The function which is now hooked at the parent of the previous one.
It doesn't change anything about how our hooks work, but makes comparing
1.16.1 and SC:R logic easier.

I think the SC:R analysis may have originally returned the same function
what we had with 1.16.1, but it is currently being inlined so the parent
function is being found instead. Updating the analysis to explicitly
find only one hook point could be preferrable..