Migrate client auth code to typescript. (#583)
This required some more complex types to get right (and to match what was already implemented, had I implemented it from scratch in TS I probably would have done it differently :) ), but this hits the goals I had, which were:

- Avoid needing to declare strings as constants that we then have to import elsewhere, and ensure they match the constant, etc.
- Allow packages to declare their own actions alongside (or inside) the interface of those actions, rather than needing one large file that contains every action
- Typecheck actions in both dispatch and reducers (and even if we dispatch a Promise that resolves to the final payload)

The reducer here is messier than other ones we have, and could probably use some cleaning up, but I left it mostly intact for the moment to avoid more structural changes than necessary. The socket-handlers here also could use some more work, as well as the interface between what we return on the server and what we fetch on the client (ideally we'd have some types that we use/share on both ends). The auth API also does a bunch of weird things and returns different parts of the user/permission info depending on how you log in, what your session is like, what client you're on, etc. and all of that could use some standardizing in the future. There's some very minor cleanup in here to fix some stuff that seemed obviously broken to me, but the rest was left as-is.