Add a manual integration of redux devtools for Electron clients. (#584)
* Add a manual integration of redux devtools for Electron clients.

Usage:
  - ctrl+h to show/hide the tools
  - ctrl+q to change position of the tools

These were the default hotkeys as specified in the walkthrough guide.
Not sure if we want to change them though?

Currently we're using the official <DockMonitor /> compnent to position
the tools, which overlays the tools over the application on the same
page. Eventually we might want to invest some time to make it display
similarly to the chrome extensions, e.g. in the same window, but not
overlaying the application, or in a popup window.

Another thing that might be useful to implement at some point is to add
a button or something which opens/closes the redux devtools, instead of
relying on hotkeys alone.

* Add types for redux-devtools-themes.

* Use the csp-nonce meta property instead of using appendChild hook to add
nonce to the devTools style element.

* Correct the development check.