Migrate matchmaking preferences actions/reducer to TS.
I discovered that this API is a bit weird and won't work properly when
we have more matchmaking types, so I'm going to clean that up in a
follow-up. The `get` portion also had matchmakingType as optional and
returned the last updated preferences in that case? But I'm not sure why
that would ever be a thing we would want so I got rid of that to reduce
the amount of stuff I have to fix. Most other stuff I left in place
until the follow-up work.