Force dateformat to 4.x.
This is used by the redux devtools, and the 3.x version is *horribly*
inefficient for their usecase (it calculates a ton of expensive formats
eagerly even though they're never actually used, for every single date
formatted, and with a lot of actions, the redux devtools format a *lot*
of dates). Haven't seen any issues with their usecase, so this seems
okay for now at least.