Remove map selection screen.
This seemed to largely be confusing to people, and I think they'll
actually be happier not "seeing how the sausage is made". We could
definitely take some steps to try to explain exactly what is happening
visually, but it's hard to do in a short time soooo... let's just try
not having that instead :) Open to further changes on this.