Add the beginnings of a party system.
This PR adds the server-side party service (and API handlers) for our
party system. Currently, only the following three actions are
implemented:
  - Send an invite to someone
  - Decline an invite (or cancel an invite you sent)
  - Accept an invite

Not sure how many more actions will be added to this PR, before we
decide it's ok to merge it. But I'm guessing at least the "Leave the
party" action as well.

NOTE: Some things will most likely change since we still haven't defined
a clear flowchart for all the possible scenarios, nor did we settle on
a UI/UX yet.