Add a page for showing match results, and APIs for retrieving them.
This is still incomplete in many ways, namely:
- Doesn't do anything with local results delivered from the game client.
  Ideally these results would be used in lieu of final results from the
  server (and the client would subscribe to + receive final results when
  they were available). The "live" indicator is a lie atm :)
- Teams are unlabeled and results are not sorted by team
- MMR change isn't shown for ranked games
- No ability to watch the game replay from this screen (this will be
  easier to handle once we store replays on the server)
- Computers are not really handled properly (they get coalesced to a
  single computer, and we never know the result). This is more of an
  issue with how game results get reported than anything.

APMs are also still all 0, but this is more of an issue with how game
results are reported.

This also doesn't navigate users here upon game completion, I'll add
that in a future commit (this work has been outstanding for quite a
while and contains a number of useful refactors/moves, so trying to get
it submitted ASAP).