Move to Jest instead of Mocha/Chai.
The maintenance story on this seemed a bit better to me, and it might
make it easier for us to add React tests down the line. At the very
least, writing new matchers is a lot easier with Jest.

Re-running tests on changes can now be accomplished with:

yarn test --watch

Once in that mode you can also specify regexes/filters for specific
tests (hit 'w' to bring up the menu).