Fix old lobby map URLs being unusable (in a hacky way).
Fixes #593.

This makes Lobbies refresh their map info whenever a new player joins.
Note that it *doesn't* notify any already-joined players of the
refreshed data, so they may still be holding an out of date URL, but for
the vast majority of cases, newly joined players will get a fresh URL
that they download immediately.

The better long-term fix for this would be to not store full map info in
the lobby system and instead provide some way to grant players access to
a map info request (such that they can temporarily access info about
another user's private maps). Our client-side system could then detect
out-of-date URL signatures in our map info cache (which we also need to
build!) and request a fresh URL when needed.

This commit also adds a user ID to our player slot structure, since it
was needed for this, and implements a way to simulate the effects of
signatures (kind of) for the local file store.