Upgrade deps. Fix some new TS errors.
The koa types are more correct now for query string params, which can
either return a string, and array of strings, or undefined. Previously
they were typed as just a string, and our code assumed as such, but if
people were to call these APIs with multiple values for those names,
they'd cause bugs/errors/etc. I added a function to deal with these
cases by taking the first value (if any values exist), thus turning the
type safely back into `string | undefined`. There are probably some
places in JS-land that we should be using this.

The TS upgrade (or maybe something pg-related?) also seems to cause
problems with pg-protocol's MessageName type, which is a const enum.
const enums *should* be fine to use in a non-exported way, but TS seems
to be unhappy with these now. `isolatedModules` was something I turned
on to match babel's "compilation" of typescript, but is the thing
causing the error here. I looked through the things we'd possibly run
into by turning it off, and the `const enum` stuff seems to be the only
thing we'd ever possibly write, so I added a babel plugin to handle them
and turned it off. (Hello future me discovering some extremely low-level
bug that wasn't caught because we had this flag off :) )