Move ActiveGameManager, GameServer, and MapStore into the main process.
I'm pretty sure Electron 12 was breaking the GameServer somehow, but it
seems like a bad idea to run an http server in the renderer process
anyway? So this gets rid of that. But to do that we needed to make it
possible to communicate with ActiveGameManager, and it touches a ton
more of the API than the stuff in the renderer process, so I chose to
move that across the boundary as well (this also gets rid of some
`remote` usage for us which is nice!).

MapStore is a dep of ActiveGameManager and we had issues with CORS in
downloading maps before anyway (I fixed them, but they were a
pain/annoying on DO Spaces), so I pushed that across the boundary as
well.

As part of this I moved some pieces of our Electron code into TS and
enabled tsyringe usage there (seemed useful and we don't care as much
about filesize in the electron app). A bunch of restructuring that could
be done to make things cleaner now that that's in place, but I'll save
that for a future date.